+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Pre release
 
+### Implemented enhancements
+- TT-1733 - Added request identity verification feature
+
 ### Fixed bugs
 - No ticket - Upgrade django to 1.11.23 to fix vulnerability